+++ tests/websocket/test_import_handler.py
@@ -29,7 +29,9 @@ from datetime import timezone
 
 import jsonschema
 import pytest
-import websockets
+import websockets.exceptions
+from websockets.asyncio.client import connect
+from websockets.asyncio.server import serve
 
 import ooresults
 from ooresults.repo.sqlite_repo import SqliteRepo
@@ -119,14 +121,15 @@ class WebSocketServer(threading.Thread):
     def run(self):
         self.loop = asyncio.new_event_loop()
         asyncio.set_event_loop(loop=self.loop)
-        self.server = self.loop.run_until_complete(
-            websockets.serve(
-                ws_handler=self.handler.handler,
-                host=self.host,
-                port=self.port,
-            ),
-        )
+        self.loop.create_task(self.start_server())
         self.loop.run_forever()
+
+    async def start_server(self) -> None:
+        self.server = await serve(
+            handler=self.handler.handler,
+            host=self.host,
+            port=self.port,
+        )
 
     async def close(self):
         self.server.close()
@@ -161,9 +164,9 @@ async def test_live_server_event_key_not_found(
   </Event>
 </ResultList>
 """
-    async with websockets.connect(
+    async with connect(
         uri="ws://localhost:8081/import",
-        extra_headers={"X-Event-Key": "xxx"},
+        additional_headers={"X-Event-Key": "xxx"},
     ) as client:
         # send a compressed IOF ResultList
         await client.send(bz2.compress(content.encode()))
@@ -173,7 +176,7 @@ async def test_live_server_event_key_not_found(
         assert response == {"result": "eventNotFound"}
 
         # websocket is closed by the server
-        with pytest.raises(websockets.ConnectionClosedOK):
+        with pytest.raises(websockets.exceptions.ConnectionClosedOK):
             await client.recv()
 
 
@@ -192,9 +195,9 @@ async def test_live_server_event_key_found_but_parse_error_in_result_list(
   </Event>
 </ResultList>
 """
-    async with websockets.connect(
+    async with connect(
         uri="ws://localhost:8081/import",
-        extra_headers={"X-Event-Key": "local"},
+        additional_headers={"X-Event-Key": "local"},
     ) as client:
         # send a compressed IOF ResultList
         await client.send(bz2.compress(content.encode()))
@@ -204,7 +207,7 @@ async def test_live_server_event_key_found_but_parse_error_in_result_list(
         assert response == {"result": "error"}
 
         # websocket is closed by the server
-        with pytest.raises(websockets.ConnectionClosedOK):
+        with pytest.raises(websockets.exceptions.ConnectionClosedOK):
             await client.recv()
 
 
@@ -224,9 +227,9 @@ async def test_live_server_event_key_found(
   </Event>
 </ResultList>
 """
-    async with websockets.connect(
+    async with connect(
         uri="ws://localhost:8081/import",
-        extra_headers={"X-Event-Key": "local"},
+        additional_headers={"X-Event-Key": "local"},
     ) as client:
         # send a compressed IOF ResultList
         await client.send(bz2.compress(content.encode()))
@@ -277,9 +280,9 @@ async def test_live_server_import_result_list_snapshot(
     s = datetime.datetime(2020, 2, 9, 10, 0, 0, tzinfo=timezone(timedelta(hours=1)))
     f = datetime.datetime(2020, 2, 9, 10, 33, 21, tzinfo=timezone(timedelta(hours=1)))
 
-    async with websockets.connect(
+    async with connect(
         uri="ws://localhost:8081/import",
-        extra_headers={"X-Event-Key": "local"},
+        additional_headers={"X-Event-Key": "local"},
     ) as client:
         # send a compressed IOF ResultList
         await client.send(bz2.compress(content.encode()))
@@ -400,9 +403,9 @@ async def test_live_server_import_result_list_delta(
     s = datetime.datetime(2020, 2, 9, 10, 0, 0, tzinfo=timezone(timedelta(hours=1)))
     f = datetime.datetime(2020, 2, 9, 10, 33, 21, tzinfo=timezone(timedelta(hours=1)))
 
-    async with websockets.connect(
+    async with connect(
         uri="ws://localhost:8081/import",
-        extra_headers={"X-Event-Key": "local"},
+        additional_headers={"X-Event-Key": "local"},
     ) as client:
         # send a compressed IOF ResultList
         await client.send(bz2.compress(content.encode()))